BUG: special.logsumexp: fix type promotion (#21528)
* BUG: special.logsumexp: fix type promotion

[skip cirrus] [skip circle]

* MAINT: adjust dtype determination logic in xp_broadcast_promote

* TST: special.logsumexp: add dtype tests

* TST: special.logsumexp: fix tests

* MAINT: define mixed-type promotion rule if underlying library doesn't

* TST: special.logsumexp: update tests

* STY: special.logsumexp: use is_array_api_strict where appropriate

[skip ci]

---------

+++ scipy/special/_logsumexp.py
@@ -129,7 +129,7 @@ def logsumexp(a, axis=None, b=None, keepdims=False, return_sign=False):
 
     # suppress warnings about log of zero
     with np.errstate(divide='ignore'):
-        s = xp.sum(tmp, axis=axis, keepdims=keepdims)
+        s = xp.sum(tmp, axis=axis, keepdims=keepdims, dtype=tmp.dtype)
         if return_sign:
             # For complex, use the numpy>=2.0 convention for sign.
             if xp.isdtype(s.dtype, "complex floating"):
